+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.5
 - Format data, but seems to be at the cost of slower processing time
+- Fix monsterSkill hit ratio display
 
 ## 0.1.4
 - Nikke setup & stat calculation